Move gloss clock rescalings to rhine-gloss
